# perf: check existsSync before calling unlinkSync