`unlinkSync` is very expensive when it throws, which it will do if `existsSync` returns `false` for the path. In my large ember app this gets hit a lot. Note that almost all of the time spent in `unlinkSync` comes from error handling, which is only needed when `unlinkSync` would throw. Adding this guard shaves about 4 seconds off of the 21 seconds spent in _copy.

Before:
`unlinkSync` takes 4.9s of 21s in `_copy`
<img width="736" alt="Screenshot 2023-10-09 at 10 13 51 PM" src="https://github.com/broccolijs/broccoli-funnel/assets/20404/d5f1d98e-b5e8-4208-b3ee-848d31918b39">


After:
`unlinkSync` takes 0s of 17s in `_copy`
<img width="852" alt="Screenshot 2023-10-09 at 10 03 09 PM" src="https://github.com/broccolijs/broccoli-funnel/assets/20404/5533c25f-db2d-4e57-8260-5494efdc2b0e">